requirements.txt
fastapi
uvicorn[standard]
easyocr
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
--index-url https://download.pytorch.org/whl/cpu
python-multipart
Pillow
opencv-python-headless
numpy